numpy2対応の積み残しとvitのvideoモードのリグレッションを修正
vit.pyのtostring_argbは4chを返すため、cv2.VideoWriterが例外を出さずに
全フレームをスキップしていた。buffer_rgba()[:, :, :3]に変更して3chに戻す。

numpy2で削除されたAPIを置換:
- np.fromstring(binary) -> np.frombuffer (mahalanobisad, psgan)
- np.product -> np.prod (patchcore)
- np.array(x, copy=False) -> np.asarray(x) (gast filterpy, ap-10k)

numpy1.24で削除済みのエイリアスを置換:
- np.long -> np.int64 (qd-3dt, damo_yolo, pytorch-dc-tts)
- np.int -> int (deepsort_vehicle)
- np.float -> float / np.float64 (riffusion)

matplotlib3.10で削除されたtostring_rgbを置換:
- uniad, bevformer, mahalanobisad -> buffer_rgba()[:, :, :3]
- 3dmppe_posenetは描画バッファへのエイリアスを切るため.copy()を追加

numpy 1.26.4 / 2.5.1 の両方で同一結果になることを確認。

+++ diffusion/riffusion/df/schedulers/scheduling_dpmsolver_multistep.py
@@ -134,8 +134,8 @@ class DPMSolverMultistepScheduler(ConfigMixin):
         dtype = sample.dtype
         batch_size, channels, height, width = sample.shape
 
-        if dtype not in (np.float, np.float64):
-            sample = sample.astype(np.float)  # upcast for quantile calculation, and clamp not implemented for cpu half
+        if dtype not in (float, np.float64):
+            sample = sample.astype(np.float64)  # upcast for quantile calculation, and clamp not implemented for cpu half
 
         # Flatten sample for doing quantile calculation along each image
         sample = sample.reshape(batch_size, channels * height * width)
